Don't regenerate the pseudo file if python or whatever is not present.

+++ autogen.sh
@@ -10,9 +10,13 @@ autoheader
 echo "Creating required files for autotools."
 automake --add-missing --copy
 echo "Generating PSP database."
-python config/pspconf.py > src/init/psp.inc
-sed '/!!PSP_TABLE!!/r src/init/psp.inc' src/init/pspconf.in.f90 > src/init/pspconf.f90
-rm -f src/init/psp.inc
+if test -f config/pspconf.py ; then
+  python config/pspconf.py > src/init/psp.inc
+  sed '/!!PSP_TABLE!!/r src/init/psp.inc' src/init/pspconf.in.f90 > src/init/pspconf.f90
+  rm -f src/init/psp.inc
+else
+  echo " WARNING, missing config/pspconf.py, cannot generate built-in pseudo-potentials."
+fi
 GDBUS_CODEGEN=`which gdbus-codegen`
 if test -n "$GDBUS_CODEGEN" ; then
   echo "Generate Dbus bindings (obsolete)."
